requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
customtkinter==5.2.1
darkdetect==0.8.0
decorator==5.1.1
future==0.18.3
geocoder==1.38.1
idna==3.6
mysql-connector==2.2.9
mysql-connector-python==8.2.0
Pillow==10.1.0
protobuf==4.21.12
pyperclip==1.8.2
ratelim==0.1.6
requests==2.31.0
six==1.16.0
tkintermapview==1.29
urllib3==2.1.0
pandas==2.1.2